% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Performs multi-tool auto-CTA}
\usage{
annotate(
  expr_data,
  ref_data,
  annot_type,
  annot_tools = "*",
  convert_to = "labels",
  tool_interfaces = list(SingleR = SingleR_interface, SCINA = SCINA_interface, ),
  ...
)
}
\arguments{
\item{expr_data}{\code{matrix} or \code{Matrix} of experimental data, on which CTA is
performed}

\item{ref_data}{\code{matrix} or \code{Matrix} of reference/marker data for the CTA
analysis}

\item{annot_type}{which type of CTA to perform <marker/ref>}

\item{annot_tools}{\code{character} vector of the names of the tools to be used.
'*' selects all of the tools in the \code{annot_type} category}

\item{convert_to}{\code{character} describing the format of the output data}

\item{tool_interfaces}{list of interfaces the program will use}

\item{labels}{\code{character} vector that holds the labels for each row of
\code{ref_data}}
}
\value{
a list of  test_name -> test_results
}
\description{
Performs multi-tool auto-CTA
}
