% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cta_interface.R
\docType{class}
\name{CTAInterface-class}
\alias{CTAInterface-class}
\title{An S4 class to represent an interface to a cell type annotation tool}
\description{
An S4 class to represent an interface to a cell type annotation tool
}
\section{Slots}{

\describe{
\item{\code{rq_list}}{\code{RequirementList} for interface parameters}

\item{\code{annotate}}{function that performs annotation}

\item{\code{convert}}{function that converts results from \code{annotate} to the format
dictated by the user}

\item{\code{preprocess_expr}}{function that transforms expression data into the
format required by an annotation tool}

\item{\code{preprocess_ref}}{fucntion that transforms reference data into the format
required by an annotation tool}
}}

