% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_processing.R
\name{find_unique_labels_per_column}
\alias{find_unique_labels_per_column}
\title{Find all of the unique values in the columns of a data.frame}
\usage{
find_unique_labels_per_column(df)
}
\arguments{
\item{df}{A data.frame to be analyzed}
}
\value{
A list of colnames -> vector of unique values
}
\description{
Find all of the unique values in the columns of a data.frame
}
