% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requirement_list.R
\docType{class}
\name{RequirementList-class}
\alias{RequirementList-class}
\title{Grouped requirements all to be applied to a single object}
\description{
Grouped requirements all to be applied to a single object
}
\section{Slots}{

\describe{
\item{\code{requirements}}{list of \code{interface_parameter} -> Requirement}
}}

