% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cta_interface.R
\name{run_full}
\alias{run_full}
\title{Runs \code{annotate_func} then \code{convert_func} on a data set}
\usage{
run_full(cta_interface, expr_data, ref_data, convert_to, ...)
}
\arguments{
\item{cta_interface}{a \code{CTAInterface} object that has implemented
\code{annotate} and \code{convert} functions for a specific tool, according to which
the annotation process will be done}

\item{expr_data}{experimental data, on which CTA is performed}

\item{ref_data}{reference/marker data for the CTA anaysis}

\item{convert_to}{output data format}
}
\value{
the annotation results from \code{cta_interface}, formatted to the
\code{convert_to} format
}
\description{
Runs \code{annotate_func} then \code{convert_func} on a data set
}
